wsd: unify UnitProxyProtocol with UnitProxy
